feat: added more custom types & printers

+++ internal/usecase/api-app/api_handler_get_signature_request.go
@@ -17,6 +17,6 @@ func (h *handler) GetSignatureRequestGet(ctx context.Context, params api.GetSign
 	return &api.SignatureRequest{
 		KeyTag:        uint8(signatureRequest.KeyTag),
 		Message:       signatureRequest.Message,
-		RequiredEpoch: signatureRequest.RequiredEpoch,
+		RequiredEpoch: uint64(signatureRequest.RequiredEpoch),
 	}, nil
 }
